Add files via upload
Added a modification to a demo sketch that was made by Ole Dreessen for the MAX77650. I added support for the IC MAX17055 and added an interupt service routine for displaying info to the MAX77650's LED controller to signal charging state. This is meant to be a demo for the PMIC (MAX77650) and the Fuel Gauge (MAX17055) the MAX32620FTHR MCU.